♻️ (#81) 검색 기능 컴포넌트에서 실행
- 검색 기능이 페이지가 아닌 SearchStar 컴포넌트에서 실행되도록 변경
